# api/rest/docs/swagger.yaml
definitions:
...
  rest.PublisherReportGetRequest:
    type: object
  rest.PublisherReportGetResponse:
    type: object
...
info:
  contact: {}
paths:
...
  /challenge/get:
    post:
      consumes:
      - application/json
      description: fetching challenges based on filter with pagination,order and selected
        fields
      parameters:
      - description: PublisherReport Get Options
        in: body
        name: options
        required: true
        schema:
          $ref: '#/definitions/rest.PublisherReportGetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.PublisherReportGetResponse'
      security:
      - ApiKeyAuth: []
      summary: Get PublisherReports.
      tags:
      - PublisherReport
...
swagger: "2.0"